+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ui/component/AppBottomBar.kt
@@ -43,6 +43,7 @@ import androidx.compose.ui.unit.sp
 import io.getstream.chat.android.compose.sample.R
 import io.getstream.chat.android.compose.ui.components.channels.UnreadCountIndicator
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import io.getstream.chat.android.ui.common.R as UiCommonR
 
 /**
  * Renders the default app bottom bar for switching between chats/threads.
@@ -78,6 +79,12 @@ fun AppBottomBar(
             },
         )
         AppBottomBarOptionTile(
+            icon = UiCommonR.drawable.stream_compose_ic_mentions,
+            text = R.string.app_bottom_bar_mentions,
+            isSelected = selectedOption == AppBottomBarOption.MENTIONS,
+            onClick = { onOptionSelected(AppBottomBarOption.MENTIONS) },
+        )
+        AppBottomBarOptionTile(
             icon = R.drawable.ic_threads,
             text = R.string.app_bottom_bar_threads,
             isSelected = selectedOption == AppBottomBarOption.THREADS,
@@ -96,6 +103,7 @@ fun AppBottomBar(
  */
 enum class AppBottomBarOption {
     CHATS,
+    MENTIONS,
     THREADS,
 }