+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -354,7 +354,7 @@ public fun MessageComposer(
 
     Surface(
         modifier = modifier,
-        shadowElevation = 4.dp,
+        shadowElevation = ChatTheme.dimens.messageComposerShadowElevation,
         color = ChatTheme.colors.barsBackground,
     ) {
         Column(Modifier.padding(vertical = 4.dp)) {