+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/list/Messages.kt
@@ -199,7 +199,7 @@ public fun Messages(
         }
     }
 
-    LoadMoreHandler(lazyListState) {
+    LoadMoreHandler(lazyListState = lazyListState) {
         if (!endOfMessages) {
             onMessagesStartReached()
         }