+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/compose/general/ChatTheme.kt
@@ -26,6 +26,7 @@ import io.getstream.chat.android.compose.ui.util.MessagePreviewFormatter
 import io.getstream.chat.android.compose.ui.util.MessageTextFormatter
 import io.getstream.chat.android.compose.ui.util.QuotedMessageTextFormatter
 import io.getstream.chat.android.compose.viewmodel.messages.MessagesViewModelFactory
+import io.getstream.chat.android.models.DraftMessage
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.ui.common.helper.DateFormatter
@@ -216,6 +217,13 @@ private object ChatThemeMessageTextFormatterDefaultSnippet : ChatThemeCustomizat
             override fun formatMessagePreview(message: Message, currentUser: User?): AnnotatedString {
                 return buildAnnotatedString {
                     append(message.text)
+                    // add your custom styling here
+                }
+            }
+
+            override fun formatDraftMessagePreview(draftMessage: DraftMessage): AnnotatedString {
+                return buildAnnotatedString {
+                    append(draftMessage.text)
                     // add your custom styling here
                 }
             }