+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/extensions/ChatClient.kt
@@ -58,12 +58,15 @@ import io.getstream.result.call.CoroutineCall
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapLatest
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.stateIn
 import java.text.SimpleDateFormat
 import java.util.Date
@@ -89,6 +92,20 @@ public val ChatClient.state: StateRegistry
 public val ChatClient.globalState: GlobalState
     @Throws(IllegalArgumentException::class)
     get() = resolveDependency<StatePlugin, GlobalState>()
+
+/**
+ * Retrieves a [Flow] holding the [GlobalState] object, which emits only if the user is connected, and the [ChatClient]
+ * is in [InitializationState.COMPLETE] state.
+ */
+public val ChatClient.globalStateFlow: Flow<GlobalState>
+    get() = clientState.initializationState
+        .onEach {
+            if (it == InitializationState.NOT_INITIALIZED) {
+                StreamLog.w(TAG) { "ChatClient::connectUser() must be called to ensure the globalState is initialized" }
+            }
+        }
+        .filter { it == InitializationState.COMPLETE }
+        .map { globalState }
 
 /**
  * [StatePluginConfig] instance used to configure [io.getstream.chat.android.state.plugin.internal.StatePlugin].