+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `DraftMessage` entity to represent a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
 
 ### ⚠️ Changed
 
@@ -17,9 +16,6 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `ChatClient.createDraftMessage()` method to create a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
-- Add `ChatClient.deleteDraftMessages()` method to delete a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
-- Add `ChatClient.queryDraftMessages()` method to query draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
 
 ### ⚠️ Changed
 
@@ -42,7 +38,6 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `GlobalState.channelDraftMessages` and `GlobalState.threadDraftMessages` properties providing access to the draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682) 
 
 ### ⚠️ Changed
 
@@ -74,13 +69,14 @@
 ## stream-chat-android-compose
 ### 🐞 Fixed
 - Fix audio recording attachments not paused when the app goes to the background or the screen is covered with another one. [#5685](https://github.com/GetStream/stream-chat-android/pull/5685)
-- Not show deleted poll messages. [#5689](https://github.com/GetStream/stream-chat-android/pull/5689)
 
 ### ⬆️ Improved
 
 ### ✅ Added
+- Introduce `MentionList` to display a list of mentions for the current user. [#5695](https://github.com/GetStream/stream-chat-android/pull/5695)
 
 ### ⚠️ Changed
+- Introduce an overload of LoadMoreHandler with threshold parameter as a lambda for calculating the value dynamically. [#5711](https://github.com/GetStream/stream-chat-android/pull/5711), [#5713](https://github.com/GetStream/stream-chat-android/pull/5713)
 
 ### ❌ Removed
 
@@ -94,6 +90,66 @@
 ### ⚠️ Changed
 
 ### ❌ Removed
+
+# April 04th, 2025 - 6.13.0
+## Common changes for all artifacts
+### 🐞 Fixed
+- Upgrade StreamPush dependency to V1.3.1 to avoid proguard issues.
+
+### ✅ Added
+- Add `DraftMessage` entity to represent a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+
+## stream-chat-android-client
+### 🐞 Fixed
+- Ensure `HeadersInterceptor` doesn't use any special character that can come from the customer integration. [#5708](https://github.com/GetStream/stream-chat-android/pull/5708)
+
+### ✅ Added
+- Add `ChatClient.createDraftMessage()` method to create a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+- Add `ChatClient.deleteDraftMessages()` method to delete a draft message. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+- Add `ChatClient.queryDraftMessages()` method to query draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682)
+
+## stream-chat-android-state
+### ✅ Added
+- Add `GlobalState.channelDraftMessages` and `GlobalState.threadDraftMessages` properties providing access to the draft messages. [#5682](https://github.com/GetStream/stream-chat-android/pull/5682) 
+- Add `ChatClient.globalStateFlow` flow holding the `GlobalState` object, which emits values only if the user is connected. [#5702](https://github.com/GetStream/stream-chat-android/pull/5702)
+
+## stream-chat-android-ui-common
+### ✅ Added
+- Introduce `MentionListController` responsible for managing the state of the mention list. [#5692](https://github.com/GetStream/stream-chat-android/pull/5692)
+
+## stream-chat-android-ui-components
+### 🐞 Fixed
+- Fix audio recording attachments not paused when the app goes to the background or the screen is covered with another one. [#5685](https://github.com/GetStream/stream-chat-android/pull/5685)
+- Fix crash happening after process death when accessing `GlobalState` from the UI components. [#5702](https://github.com/GetStream/stream-chat-android/pull/5702)
+
+### ⬆️ Improved
+- Enable pagination in `MentionListView`. [#5692](https://github.com/GetStream/stream-chat-android/pull/5692)
+
+### ✅ Added
+- Add `ChatUI.draftMessagesEnabled` property to enable/disable Draft Messages. [#5687](https://github.com/GetStream/stream-chat-android/pull/5687)
+
+### ⚠️ Changed
+- 🚨Breaking change: Move `MentionListViewModel` logic and its state to a shared component so they can be reused in Compose. [#5692](https://github.com/GetStream/stream-chat-android/pull/5692)
+- 🚨Breaking change: `ChannelListViewModel` now accepts a `Flow<GlobalState>` instead of `GlobalState` for the `globalState` constructor parameter.  [#5702](https://github.com/GetStream/stream-chat-android/pull/5702)
+
+## stream-chat-android-compose
+### 🐞 Fixed
+- Fix audio recording attachments not paused when the app goes to the background or the screen is covered with another one. [#5685](https://github.com/GetStream/stream-chat-android/pull/5685)
+- Not show deleted poll messages. [#5689](https://github.com/GetStream/stream-chat-android/pull/5689)
+- Fix "Thread reply" item shown in the message options menu for messages in a Thread. [#5683](https://github.com/GetStream/stream-chat-android/pull/5683)
+- Fix crash happening after process death when accessing `GlobalState` from the UI components. [#5702](https://github.com/GetStream/stream-chat-android/pull/5702)
+
+### ✅ Added
+- Add `MessagesViewModelFactory.isComposerDraftMessageEnabled` property to enable/disable Draft Messages within `MessageComposer`. [#5687](https://github.com/GetStream/stream-chat-android/pull/5687)
+- Add `ChannelViewModelFactory.isDraftMessageEnabled` property to enable/disable Draft Messages within `ChannelList`. [#5687](https://github.com/GetStream/stream-chat-android/pull/5687)
+- Add `MessageBubble` and `MessageFailedIcon` to `ChatComponentFactory` to allow customization of the message bubbles and the failed message icon. [#5688](https://github.com/GetStream/stream-chat-android/pull/5688)
+- Add `errorTextStyle` and `errorBackgroundColor` to `MessageTheme` to allow customization of the error message bubbles and error message text style. [#5688](https://github.com/GetStream/stream-chat-android/pull/5688)
+- ⚠️Add `Message` as an argument of the `textStyle` lambda parameter of the `MessageTextFormatter.defaultFormatter` method to allow customizing the text style based on the `Message`. [#5688](https://github.com/GetStream/stream-chat-android/pull/5688)
+- Create `ChatTheme.dimens.messageComposerShadowElevation` to customize the message composer shadow elevation. [#5697](https://github.com/GetStream/stream-chat-android/pull/5697)
+
+### ⚠️ Changed
+- `defaultMessageOptionsState()` now accepts an `isInThread` flag to show/hide the "Thread reply" option. [#5683](https://github.com/GetStream/stream-chat-android/pull/5683)
+- Make `onSearchResultClick` optional in `SearchResultItem`, so we don't render a ripple effect if it is not clickable. [#5695](https://github.com/GetStream/stream-chat-android/pull/5695)
 
 # March 14th, 2025 - 6.12.1
 ## Common changes for all artifacts